ImageCompressionFunction/requirements.txt
azure-functions
Pillow
pytest